style: improve structure of clock code
