+++ src/ui/interface.rs
@@ -1,39 +1,12 @@
-use crate::{
-    ui::{self, components, window::Window},
-    Args,
-};
+use crate::{ui, Args};
 use std::{env, time::Duration};
-use tokio::time::Instant;
 
-/// An extremely simple clock to be used alongside the [`Window`].
-pub struct Clock(Instant);
+pub mod clock;
+pub mod components;
+pub mod window;
 
-impl Clock {
-    /// Small shorthand for getting the local time now, and formatting it.
-    #[inline]
-    fn now() -> chrono::format::DelayedFormat<chrono::format::StrftimeItems<'static>> {
-        chrono::Local::now().format("%H:%M:%S")
-    }
-
-    /// Checks if the last update was long enough ago, and if so,
-    /// updates the displayed clock.
-    ///
-    /// This is to avoid constant calls to [`chrono::Local::now`], which
-    /// is somewhat expensive because of timezones.
-    pub fn update(&mut self, window: &mut Window) {
-        if self.0.elapsed().as_millis() >= 500 {
-            window.display(Self::now(), 8);
-            self.0 = Instant::now();
-        }
-    }
-
-    /// Simply creates a new clock, and renders it's initial state to the window top.
-    pub fn new(window: &mut Window) -> Self {
-        window.display(Self::now(), 8);
-
-        Self(Instant::now())
-    }
-}
+pub use clock::Clock;
+pub use window::Window;
 
 /// UI-specific parameters and options.
 #[derive(Copy, Clone, Debug, Default)]
@@ -47,7 +20,7 @@ pub struct Params {
     /// Whether the visual part of the UI should be enabled.
     /// This only applies if the MPRIS feature is enabled.
     pub enabled: bool,
-  
+
     /// Whether to include the clock on the top bar.
     pub clock: bool,
